refactor(wizards/connectedap): move wizards
